+++ src/front/interface/aceditor.js
@@ -4,10 +4,10 @@
 
 const $ = require('jquery');
 let ace = require('ace-builds/src-min/ace');
-const Languages = require('./languages');
+const Languages = require('./../languages');
 
 let aceditor = null;
-const langInfo = {};
+// const langInfo = {};
 
 // _____________________________________________________
 // initialize
