+++ src/front/interface/panel.js
