+++ src/front/interface/stdios.js
